refactor: UI 와 기능 (훅)으로 분리
